Check current_node not nil in open()
Opening the context window unconditionally via open() would cause errors
if there was nothing to show (current_node == nil). This could be
triggered by the User autocmd for SaveSessionPost.

This could instead be fixed by using update_context() in
SaveSessionPost, but there shouldn't be any need to re-compute the
context in this case.