Add highlight group for customization
Currently the highlight of the floating window to show the context is
fixed by NeoVims default to 'NormalFloat'. The user has no option to
customize it selectively without changing the appearance for all
floating windows.
Therefore a new highlight group gets added, that still defaults to the
'NormalFloat' group. Though the user can now set this group to change
the plugins colors.