fix: schedule buffer update
Any buffer should only be updated when it is safe. Therefore, to avoid
textlock or other temporary restriction, the function call should be
scheduled in the main event-loop by using `vim.schedule()`

This fixes visual anomalies that occur when scrolling all the way to the
end of a page when a context is open.