test: properly check treesitter lang and only install parser when needed
Because of the way we were checking the treesitter lang, we were relying
on the parser to already be installed for the test to not be skipped.
Now we check it via an `nvim-treesitter` API that does not require the
parser to be installed. This results in all the parsers to be installed
even when the related tests won't be run, so we do a lazy setup here as
well.