test: support injected query tests
Before we had snapshots to cover injected query tests, but now we can
use the cursor and context markers to test these as well. This will mean
that we might be installing more parsers and parsing more than we should
be, but I think the higher test coverage is worth it.